Renamed index route file to terms
